+++ src/server/metrics/metrics.js
@@ -1,0 +1,3 @@
+export const UserLogin = 'METRIC_USER_LOGIN';
+export const ProxyRequest = 'METRIC_PROXY_REQUEST';
+export const InboundRequest = 'METRIC_INBOUND_REQUEST';